pip-requirements.txt
setuptools
scikit-learn
nltk
gensim
angus-sdk-python
pymongo
tweepy
lxml
Flask
Flask-Script